+++ tests/PSRule.Tests/FromFileInvalid.Rule.ps1
@@ -1,0 +1,10 @@
+
+# Synopsis: Null DependsOn is invalid.
+Rule 'InvalidRule1' -DependsOn $Null {
+
+}
+
+# Synopsis: Empty DependsOn collection is invalid.
+Rule 'InvalidRule2' -DependsOn @($Null) {
+
+}